python-scrapper/requirements.txt
mysql.connector==2.2.9
bs4==*